Make Clap sort subcommands automatically
By default `clap` will use the order subcommands are added to a
`Command` to determine print order in `--help`. In our case this results
in the help text being randomly sorted, which is a confusing experience
for users.

Add `.display_order(0)` when creating a new `Command` so that `clap`
will automatically sort the subcommands for us.