Return source of wrapped errors
The `Display` impl for `Error` wraps the original error received with a
message, e.g., "Communication error: {original_error_message}". The
`source` method returns the same error, though, so iterating over the
error chain will cause the same error message to be printed twice.

Update `source` to return the source error of the original error, not
our wrapper, to avoid this duplication.