Add option to select HTTP client type
Currently `reqwest::Client` is the only type we allow as the underlying
HTTP client. Some consumers of Progenitor may want to use
`reqwest_middleware` to enable conveniences like automatic retries and
support for `tracing`.

Add a new `with_client_type` argument for `GenerationSettings` to allow
callers to select which of these types to use as the client.